+++ packages/bot/utils/interactive.js
@@ -4,9 +4,7 @@ const printer = (message, title) => {
     if (NODE_ENV !== 'test') {
         // console.clear()
         if (title) console.log(bgRed(`${title}`))
-        console.log(
-            yellow(Array.isArray(message) ? message.join('\n') : message)
-        )
+        console.log(yellow(Array.isArray(message) ? message.join('\n') : message))
         console.log(``)
     }
 }